delete DeviceServer.js and main.js / rename DeviceServer_v2.js
