+++ src/index.js
@@ -21,7 +21,7 @@
 import DeviceAttributeFileRepository from './repository/DeviceAttributeFileRepository';
 import DeviceKeyFileRepository from './repository/DeviceKeyFileRepository';
 import EventPublisher from './lib/EventPublisher';
-import DeviceServer from './server/DeviceServer_v2';
+import DeviceServer from './server/DeviceServer';
 import FileManager from './repository/FileManager';
 import JSONFileManager from './repository/JSONFileManager';
 import ServerConfigFileRepository from './repository/ServerConfigFileRepository';
